+++ client/src/app/routes/todos/todo-edit-dialog/todo-edit-dialog.component.ts
@@ -1,0 +1,11 @@
+import { Component } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+@Component({
+  selector: 'app-todo-edit-dialog',
+  standalone: true,
+  imports: [CommonModule],
+  templateUrl: './todo-edit-dialog.component.html',
+  styleUrl: './todo-edit-dialog.component.scss',
+})
+export class TodoEditDialogComponent {}